Enable opting-out of form submissions
Similar to [Disabling Turbo Drive on Specific Links][turbo-drive-false],
guard intercepting form subissions with a similar check for
`[data-turbo-drive="false"]` on the submitter, `<form>` element, or an
ancestor.

Similarly, *do not intercept* when the submitter or form element
specifies a method that isn't an HTTP verb (via
[formmethod="dialog"][formmethod] attribute or [method="dialog"][method]
attribute, respectively).

In practice, `<form>` elements will only ever declare `method="get"` or
`method="post"`, but given the addition of `method="dialog"` and the
fact that the `fetchMethodFromString` and the `FetchMethod` enum type
exist, we might as well be careful and future-proof.

[turbo-drive-false]: https://turbo.hotwire.dev/handbook/drive#disabling-turbo-drive-on-specific-links
[formmethod]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-formmethod
[method]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form#attr-method